Move main functions to vdom module
