+++ tests/core/benchmark/app/Row.ts
@@ -1,6 +1,6 @@
 import { DNode } from '../../../../src/core/interfaces';
 import { WidgetBase } from '../../../../src/core/WidgetBase';
-import { v } from '../../../../src/core/d';
+import { v } from '../../../../src/core/vdom';
 
 export interface RowProperties {
 	onRowSelected: (id: number) => void;
