+++ tests/stores/unit/StoreContainer.ts
@@ -1,7 +1,7 @@
 const { beforeEach, describe, it } = intern.getInterface('bdd');
 const { assert } = intern.getPlugin('chai');
 import { WidgetBase } from '../../../src/core/WidgetBase';
-import { v, WNODE } from '../../../src/core/d';
+import { v } from '../../../src/core/vdom';
 import { Registry } from '../../../src/core/Registry';
 
 import DefaultStoreContainer, { createStoreContainer, StoreContainer } from '../../../src/stores/StoreContainer';
@@ -43,7 +43,7 @@ describe('StoreContainer', () => {
 			assert.deepEqual(renderResult, {
 				properties: { key: '1' },
 				children: [child],
-				type: WNODE,
+				type: '__WNODE_TYPE',
 				widgetConstructor: Foo
 			});
 		});
@@ -93,7 +93,7 @@ describe('StoreContainer', () => {
 			assert.deepEqual(renderResult, {
 				properties: { key: '1' },
 				children: [child],
-				type: WNODE,
+				type: '__WNODE_TYPE',
 				widgetConstructor: Foo
 			});
 		});
@@ -144,7 +144,7 @@ describe('StoreContainer', () => {
 			assert.deepEqual(renderResult, {
 				properties: { key: '1' },
 				children: [child],
-				type: WNODE,
+				type: '__WNODE_TYPE',
 				widgetConstructor: Foo
 			});
 		});
