+++ tests/stores/unit/StoreInjector.ts
@@ -2,7 +2,7 @@ const { beforeEach, describe, it } = intern.getInterface('bdd');
 const { assert } = intern.getPlugin('chai');
 import { spy } from 'sinon';
 import { WidgetBase } from '../../../src/core/WidgetBase';
-import { v, WNODE } from '../../../src/core/d';
+import { v } from '../../../src/core/vdom';
 import { Registry } from '../../../src/core/Registry';
 
 import {
@@ -293,7 +293,7 @@ describe('StoreInjector', () => {
 			assert.deepEqual(renderResult, {
 				properties: { key: '1' },
 				children: [child],
-				type: WNODE,
+				type: '__WNODE_TYPE',
 				widgetConstructor: Foo
 			});
 		});
@@ -343,7 +343,7 @@ describe('StoreInjector', () => {
 			assert.deepEqual(renderResult, {
 				properties: { key: '1' },
 				children: [child],
-				type: WNODE,
+				type: '__WNODE_TYPE',
 				widgetConstructor: Foo
 			});
 		});
