+++ tests/testing/unit/harness.ts
@@ -3,7 +3,7 @@ const { assert } = intern.getPlugin('chai');
 
 import { harness } from '../../../src/testing/harness';
 import { WidgetBase } from '../../../src/core/WidgetBase';
-import { v, w, isVNode } from '../../../src/core/d';
+import { v, w, isVNode } from '../../../src/core/vdom';
 import Set from '../../../src/shim/Set';
 import Map from '../../../src/shim/Map';
 import { VNode, WNode } from '../../../src/core/interfaces';
