+++ tests/testing/unit/harnessWithTsx.tsx
@@ -3,7 +3,7 @@ const { assert } = intern.getPlugin('chai');
 
 import { harness } from '../../../src/testing/harness';
 import { WidgetBase } from '../../../src/core/WidgetBase';
-import { tsx, fromRegistry } from '../../../src/core/tsx';
+import { tsx, fromRegistry } from '../../../src/core/vdom';
 
 class ChildWidget extends WidgetBase<any> {}
 const RegistryWidget = fromRegistry<any>('registry-item');
